+++ IncomeExpensedetails.java
@@ -1,5 +1,6 @@
 package pft;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
 import java.util.ArrayList;
@@ -10,13 +11,10 @@ import java.util.Scanner;
 
 public interface IncomeExpensedetails<T> {
 
-   
- 
-
-    void addIncome(T income,double amount, LocalDate date);
+    void addIncome(T income, BigDecimal amount, LocalDate date);
     void displayIncomeSources();
     void deleteIncome(int id);
-    void updateIncome(int id, T newIncome, double newAmount, LocalDate newDate);
+    void updateIncome(int id, T newIncome, BigDecimal newAmount, LocalDate newDate);
     void findIncomeById(int id);
     void findParticularColumnById(int id,String column);
     void showMenu();
